using double instead of float because EPiServer not support System.Single

+++ Jumeirah.DummyData.Cms/BlockDatas/HotelBlock.cs
@@ -67,7 +67,8 @@ namespace Jumeirah.DummyData.Cms.BlockDatas
         public virtual string LocalCurrency { get; set; }
 
         [Display(GroupName = Constants.ExtendTab, Order = 10)]
-        public virtual float PriceUSD { get; set; }
+        public virtual double PriceUSD { get; set; } 
+        //public virtual float PriceUSD { get; set; } // Type 'System.Single' could not be mapped to a PropertyDefinitionType
 
         [Display(GroupName = Constants.ExtendTab, Order = 11)]
         public virtual string PriceFrom { get; set; }
